@@ -20,20 +20,30 @@
     },
     "require": {
         "php": ">=5.4.0",
-        "psr/log": "~1.0",
-        "symfony/console": "~2.5|~3.0",
-        "symfony/finder": "~2.5|~3.0",
-        "phpdocumentor/reflection-docblock": "~2"
+        "consolidation/output-formatters": "^3.1.5",
+        "psr/log": "~1",
+        "symfony/console": "^2.8|~3",
+        "symfony/event-dispatcher": "^2.5|~3",
+        "symfony/finder": "^2.5|~3",
+        "phpdocumentor/reflection-docblock": "^2.0|^3.0.2"
     },
     "require-dev": {
-        "consolidation/output-formatters": "~1",
         "phpunit/phpunit": "4.*",
         "satooshi/php-coveralls": "^1.0",
-        "squizlabs/php_codesniffer": "2.*"
+        "squizlabs/php_codesniffer": "^2.7"
+    },
+    "scripts": {
+        "cs": "phpcs --standard=PSR2 -n src",
+        "cbf": "phpcbf --standard=PSR2 -n src",
+        "unit": "SHELL_INTERACTIVE=true phpunit --colors=always",
+        "test": [
+            "@unit",
+            "@cs"
+        ]
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.x-dev"
+            "dev-master": "2.x-dev"
         }
     }
 }
