@@ -2,46 +2,44 @@
 
 namespace Drupal\sharemessage\Tests;
 
-use Drupal\file\Entity\File;
-
 /**
- * Check if default and Share Message specific settings work correctly.
+ * Check if default Share Message settings work correctly.
  *
  * @group sharemessage
  */
 class ShareMessageSettingsTest extends ShareMessageTestBase {
 
   /**
-   * Test case that check if default and Share Message specific settings work correctly.
+   * Tests if default and Share Message specific settings work correctly.
    */
   public function testShareMessageSettings() {
 
     // Step 1: Setup default settings.
     $this->drupalGet('admin/config/services/sharemessage/addthis-settings');
-    $default_settings = array(
-      'default_services[]' => array(
+    $default_settings = [
+      'default_services[]' => [
         'facebook',
         'facebook_like',
-      ),
+      ],
       'default_additional_services' => FALSE,
       'default_icon_style' => 'addthis_16x16_style',
-    );
+    ];
     $this->drupalPostForm(NULL, $default_settings, t('Save configuration'));
 
     // Step 2: Create Share Message with customized settings.
     $this->drupalGet('admin/config/services/sharemessage/add');
-    $sharemessage = array(
+    $sharemessage = [
       'label' => 'Share Message Test Label',
       'id' => 'sharemessage_test_label',
       'settings[override_default_settings]' => 1,
-      'settings[services][]' => array(
+      'settings[services][]' => [
         'facebook',
-      ),
+      ],
       'settings[additional_services]' => 1,
       'settings[icon_style]' => 'addthis_32x32_style',
-    );
+    ];
     $this->drupalPostForm(NULL, $sharemessage, t('Save'));
-    $this->assertText(t('Share Message @label has been added.', array('@label' => $sharemessage['label'])));
+    $this->assertText(t('Share Message @label has been added.', ['@label' => $sharemessage['label']]));
 
     // Step 3: Verify that settings are overridden
     // (services, additional_services and icon_style).
@@ -60,11 +58,11 @@ class ShareMessageSettingsTest extends ShareMessageTestBase {
 
     // Step 4: Uncheck "Override default settings" checkbox.
     $this->drupalGet('admin/config/services/sharemessage/manage/' . $sharemessage['id']);
-    $edit = array(
+    $edit = [
       'settings[override_default_settings]' => FALSE,
-    );
+    ];
     $this->drupalPostForm(NULL, $edit, t('Save'));
-    $this->assertText(t('Share Message @label has been updated.', array('@label' => $sharemessage['label'])));
+    $this->assertText(t('Share Message @label has been updated.', ['@label' => $sharemessage['label']]));
 
     // Step 5: Check that addThis widget is displayed with default settings.
     $this->drupalGet('sharemessage-test/sharemessage_test_label');
@@ -86,12 +84,12 @@ class ShareMessageSettingsTest extends ShareMessageTestBase {
     $this->assertNoLinkByHref('admin/config/services/sharemessage/sharrre-settings');
 
     // Test for empty remote library URL.
-    $settings_without_url = array(
-      'default_services[]' => array(
+    $settings_without_url = [
+      'default_services[]' => [
         'googlePlus',
-      ),
+      ],
       'library_url' => '',
-    );
+    ];
     $this->drupalPostForm(NULL, $settings_without_url, t('Save configuration'));
     $this->assertText('The configuration options have been saved.');
     $this->assertText('Either set the library locally (in /libraries/sharrre) and enable the libraries module or enter the remote URL on Sharrre settings page.');
@@ -99,43 +97,43 @@ class ShareMessageSettingsTest extends ShareMessageTestBase {
 
     // Test for wrong naming of remote library.
     $this->drupalGet('admin/config/services/sharemessage/sharrre-settings');
-    $settings_with_wrong_url = array(
-      'default_services[]' => array(
+    $settings_with_wrong_url = [
+      'default_services[]' => [
         'googlePlus',
-      ),
+      ],
       'library_url' => 'test/sharrre.js',
-    );
+    ];
     $this->drupalPostForm(NULL, $settings_with_wrong_url, t('Save configuration'));
     $this->assertText('The configuration options have been saved.');
     $this->assertText('The remote URL is unexpected. Please, provide the correct URL to the minimized version of the library found on Sharrre CDN.');
 
-    $default_settings = array(
-      'default_services[]' => array(
+    $default_settings = [
+      'default_services[]' => [
         'googlePlus',
         'facebook',
         'twitter',
         'linkedin',
         'pinterest',
-      ),
+      ],
       'library_url' => '//cdn.jsdelivr.net/sharrre/1.3.4/jquery.sharrre-1.3.4.min.js',
       'shorter_total' => TRUE,
       'enable_hover' => TRUE,
       'enable_counter' => FALSE,
-    );
+    ];
     $this->drupalPostForm(NULL, $default_settings, t('Save configuration'));
 
     // Step 2: Switch to Sharrre plugin.
     $this->drupalGet('admin/config/services/sharemessage');
     $this->clickLink('Edit');
-    $sharemessage = array(
+    $sharemessage = [
       'label' => 'Share Message Sharrre Test Label',
       'plugin' => 'sharrre',
       'title' => 'Sharrre',
       'message_long' => 'Test long message',
       'message_short' => 'Test short message',
-    );
+    ];
     $this->drupalPostForm(NULL, $sharemessage, t('Save'));
-    $this->assertText(t('Share Message @label has been updated.', array('@label' => $sharemessage['label'])));
+    $this->assertText(t('Share Message @label has been updated.', ['@label' => $sharemessage['label']]));
 
     $this->drupalGet('sharemessage-test/sharemessage_test_label');
     $this->assertRaw('"library_url":"\/\/cdn.jsdelivr.net\/sharrre\/1.3.4\/jquery.sharrre-1.3.4.min.js"');
@@ -143,23 +141,23 @@ class ShareMessageSettingsTest extends ShareMessageTestBase {
 
     // Test the naming of the file warning.
     $this->drupalGet('admin/config/services/sharemessage/sharrre-settings');
-    $settings_with_wrong_library_naming = array(
-      'default_services[]' => array(
+    $settings_with_wrong_library_naming = [
+      'default_services[]' => [
         'facebook'
-      ),
+      ],
       'library_url' => '//cdn.jsdelivr.net/sharrre/1.3.4/jquery.sharrre.js',
-    );
+    ];
     $this->drupalPostForm(NULL, $settings_with_wrong_library_naming, t('Save configuration'));
     $this->assertText('The configuration options have been saved.');
     $this->assertText('The naming of the library is unexpected. Double check that this is the real Sharrre library. The URL for the minimized version of the library can be found on Sharrre CDN.');
 
     // Test if preg match for the naming of the library works correctly.
-    $settings_with_correct_library_naming = array(
-      'default_services[]' => array(
+    $settings_with_correct_library_naming = [
+      'default_services[]' => [
         'facebook'
-      ),
+      ],
       'library_url' => '//cdn.jsdelivr.net/sharrre/1.3.4/jquery.sharrre-10.130.1234.min.js',
-    );
+    ];
     $this->drupalPostForm(NULL, $settings_with_correct_library_naming, t('Save configuration'));
     $this->assertText('The configuration options have been saved.');
     $this->assertNoText('The naming of the library is unexpected. Double check that this is the real Sharrre library. The URL for the minimized version of the library can be found on Sharrre CDN.');
@@ -167,16 +165,69 @@ class ShareMessageSettingsTest extends ShareMessageTestBase {
     // Test UrlCurl script for Sharrre.
     // This test requires capability to connect to Google Plus, Stumbleupon,
     // Pinterest platforms.
-    $this->drupalGet('/sharemessage/sharrre/counter', array('query' => array('url' => 'https://www.drupal.org/', 'type' => 'googlePlus')));
+    $this->drupalGet('/sharemessage/sharrre/counter', ['query' => ['url' => 'https://www.drupal.org/', 'type' => 'googlePlus']]);
     $json_content = json_decode($this->getRawContent(), TRUE);
     $this->assertTrue(isset($json_content['count']), 'googlePlus count found and has a non-zero value: ' . $json_content['count']);
-    $this->drupalGet('/sharemessage/sharrre/counter', array('query' => array('url' => 'https://www.drupal.org/', 'type' => 'stumbleupon')));
+    $this->drupalGet('/sharemessage/sharrre/counter', ['query' => ['url' => 'https://www.drupal.org/', 'type' => 'stumbleupon']]);
     $json_content = json_decode($this->getRawContent(), TRUE);
     $this->assertTrue(isset($json_content['count']), 'stumbleupon count found and has a non-zero value: ' . $json_content['count']);
+
+    // Test Social Share Privacy.
+    $this->drupalGet('admin/config/services/sharemessage/socialshareprivacy-settings');
+    // Step 1: Set up default settings for Social Share Privacy.
+    $default_settings = [
+      'services[]' => [
+        'gplus',
+        'twitter',
+      ],
+    ];
+    $this->drupalPostForm(NULL, $default_settings, t('Save configuration'));
+
+    $this->drupalGet('admin/config/services/sharemessage/add');
+    // Step 2: Create Social Share Privacy with customized settings.
+    $sharemessage = [
+      'label' => 'Social Share Privacy Test Label',
+      'id' => 'socialshareprivacy_test_label',
+      'settings[override_default_settings]' => 1,
+      'settings[services][]' => [
+        'facebook',
+      ],
+      'plugin' => 'socialshareprivacy',
+    ];
+    $this->drupalPostForm(NULL, $sharemessage, t('Save'));
+    $this->assertText(t('Share Message @label has been added.', ['@label' => $sharemessage['label']]));
+
+    // Step 3: Verify that settings are overridden.
+    $this->drupalGet('sharemessage-test/socialshareprivacy_test_label');
+    $this->assertRaw('"facebook":{"status":true');
+    // Step 4: Uncheck "Override default settings" checkbox.
+    $this->drupalGet('admin/config/services/sharemessage/manage/' . $sharemessage['id']);
+    $edit = [
+      'settings[override_default_settings]' => FALSE,
+    ];
+    $this->drupalPostForm(NULL, $edit, t('Save'));
+    $this->assertText(t('Share Message @label has been updated.', ['@label' => $sharemessage['label']]));
+    // Step 5: Check default settings of Social Share Privacy is displayed.
+    $this->drupalGet('sharemessage-test/socialshareprivacy_test_label');
+    // Check services (googlePlus and twitter should be displayed).
+    $this->assertRaw('"gplus":{"status":true');
+    $this->assertRaw('"twitter":{"status":true');
+    $this->assertRaw('"facebook":{"status":false');
+
+    // Test sharemessage global settings.
+    $this->drupalGet('admin/config/services/sharemessage/sharemessage-settings');
+    $edit = [
+      'add_twitter_card' => TRUE,
+      'twitter_user' => 'fancy_twitter_name',
+    ];
+    $this->drupalPostForm(NULL, $edit, t('Save configuration'));
+    $this->assertText(t('The configuration options have been saved.'));
+    $this->assertFieldByName('add_twitter_card', TRUE);
+    $this->assertFieldByName('twitter_user', 'fancy_twitter_name');
   }
 
   /**
-   * Test case for the delete and cancel button functionality for Share Message.
+   * Tests the Share Message delete and cancel button functionality.
    */
   function testShareMessageDeleteCancel() {
     // Create a Share Message.
@@ -218,53 +269,21 @@ class ShareMessageSettingsTest extends ShareMessageTestBase {
   }
 
   /**
-   * Test case for OGHeaders plugin.
-   */
-  public function testShareMessageOGTags() {
-    // Create Share Message with OG headers as plugin.
-    $this->drupalGet('admin/config/services/sharemessage/add');
-    file_put_contents('public://file.png', str_repeat('t', 8000000));
-    $file = File::create(['uri' => 'public://file.png']);
-    $file->save();
-    $sharemessage = array(
-      'label' => 'Share Message Test OG Label',
-      'id' => 'sharemessage_test_og_label',
-      'plugin' => 'ogheaders',
-      'title' => 'OG headers name',
-      'message_long' => 'OG headers long description',
-      'message_short' => 'OG headers short description',
-      'fallback_image' => $file->uuid(),
-    );
-    $this->drupalPostForm(NULL, $sharemessage, t('Save'));
-    $this->assertText(t('Share Message @label has been added.', array('@label' => $sharemessage['label'])));
-    $this->drupalGet('sharemessage-test/sharemessage_test_og_label');
-    $url = file_create_url($file->getFileUri());
-
-    // Check for OG headers in meta tags.
-    $this->assertOGTags('og:title', 'OG headers name');
-    $this->assertOGTags('og:url', $this->url);
-    $this->assertOGTags('og:description', 'OG headers long description');
-    $this->assertOGTags('og:image', $url);
-
-    // @todo Add coverage for special characters in og tags.
-  }
-
-  /**
    * Test case for special characters encoding.
    */
   public function testShareMessageSpecialCharsEncoding() {
-    // Create Share Message with OG headers as plugin.
+    // Create Share Message with AddThis as plugin.
     $this->drupalGet('admin/config/services/sharemessage/add');
-    $sharemessage = array(
+    $sharemessage = [
       'label' => 'Special characters encoding test',
       'id' => 'sharemessage_test_special_chars',
       'plugin' => 'addthis',
       'title' => 'Inondations sur la Côte d\'Azur: «C’est apocalyptique, c’est Tchernobyl»',
       'message_long' => 'Long description',
       'message_short' => 'Short description',
-    );
+    ];
     $this->drupalPostForm(NULL, $sharemessage, t('Save'));
-    $this->assertText(t('Share Message @label has been added.', array('@label' => $sharemessage['label'])));
+    $this->assertText(t('Share Message @label has been added.', ['@label' => $sharemessage['label']]));
     $this->drupalGet('sharemessage-test/sharemessage_test_special_chars');
     // Check for correct encoding in meta tags.
     $this->assertOGTags('og:title', 'Inondations sur la Côte d&#039;Azur: «C’est apocalyptique, c’est Tchernobyl»');
@@ -274,134 +293,18 @@ class ShareMessageSettingsTest extends ShareMessageTestBase {
     $this->assertOGTags('og:type', 'website');
 
     $this->drupalGet('admin/config/services/sharemessage/add');
-    $sharemessage2 = array(
+    $sharemessage2 = [
       'label' => 'Special characters encoding test 2',
       'id' => 'sharemessage_test_special_chars_2',
       'plugin' => 'addthis',
       'title' => 'This is a second test with quotes "',
       'message_long' => 'Long description 2',
       'message_short' => 'Short description 2',
-    );
+    ];
     $this->drupalPostForm(NULL, $sharemessage2, t('Save'));
     $this->drupalGet('sharemessage-test/sharemessage_test_special_chars_2');
     $this->assertOGTags('og:title', 'This is a second test with quotes &quot;');
     $this->assertNoOGTags('og:title', 'This is a second test with quotes "');
-  }
-
-  /**
-   * Test case for Sharrre settings form saving.
-   */
-  public function testSharrreSettingsFormSave() {
-    // Set initial Sharrre settings.
-    $this->drupalGet('admin/config/services/sharemessage/sharrre-settings');
-    $default_settings = array(
-      'default_services[]' => array(
-        'googlePlus',
-        'facebook',
-      ),
-      'library_url' => '//cdn.jsdelivr.net/sharrre/1.3.4/jquery.sharrre-1.3.4.min.js',
-      'shorter_total' => FALSE,
-      'enable_hover' => FALSE,
-      'enable_counter' => FALSE,
-      'enable_tracking' => FALSE,
-    );
-    $this->drupalPostForm(NULL, $default_settings, t('Save configuration'));
-
-    // Set a new Share Message.
-    $this->drupalGet('admin/config/services/sharemessage/add');
-    $sharemessage = array(
-      'label' => 'ShareMessage Test Sharrre',
-      'id' => 'sharemessage_test_sharrre_label',
-      'plugin' => 'sharrre',
-      'title' => 'Sharrre test',
-    );
-    $this->drupalPostForm(NULL, $sharemessage, t('Save'));
-
-    // Assert that the initial settings are saved correctly.
-    $this->drupalGet('sharemessage-test/sharemessage_test_sharrre_label');
-    $this->assertRaw('"services":{"googlePlus":"googlePlus","facebook":"facebook"}');
-    $this->assertRaw('"shorter_total":false,"enable_hover":false,"enable_counter":false,"enable_tracking":false');
-
-    // Set new Sharrre settings.
-    $this->drupalGet('admin/config/services/sharemessage/sharrre-settings');
-    $default_settings = array(
-      'default_services[]' => array(
-        'googlePlus',
-        'facebook',
-        'twitter',
-        'linkedin',
-        'pinterest',
-      ),
-      'library_url' => '//cdn.jsdelivr.net/sharrre/1.3.4/jquery.sharrre-1.3.4.min.js',
-      'shorter_total' => TRUE,
-      'enable_hover' => TRUE,
-      'enable_counter' => TRUE,
-      'enable_tracking' => FALSE,
-    );
-    $this->drupalPostForm(NULL, $default_settings, t('Save configuration'));
-
-    // Check that the saving of the new Sharrre settings works correctly.
-    $this->drupalGet('sharemessage-test/sharemessage_test_sharrre_label');
-    $this->assertRaw('"services":{"googlePlus":"googlePlus","facebook":"facebook","twitter":"twitter","linkedin":"linkedin","pinterest":"pinterest"}');
-    $this->assertNoRaw('"services":{"googlePlus":"googlePlus","facebook":"facebook"}');
-    $this->assertRaw('"shorter_total":true,"enable_hover":true,"enable_counter":true,"enable_tracking":false');
-    $this->assertNoRaw('"shorter_total":false,"enable_hover":false,"enable_counter":false,"enable_tracking":false');
-  }
-
-  /**
-   * Test case for AddThis settings form saving.
-   */
-  public function testAddThisSettingsFormSave() {
-    // Set initial AddThis settings.
-    $this->drupalGet('admin/config/services/sharemessage/addthis-settings');
-    $default_settings = array(
-      'default_services[]' => array(
-        'facebook',
-        'facebook_like',
-      ),
-      'default_additional_services' => FALSE,
-      'default_icon_style' => 'addthis_16x16_style',
-    );
-    $this->drupalPostForm(NULL, $default_settings, t('Save configuration'));
-
-    // Set a new Share Message.
-    $this->drupalGet('admin/config/services/sharemessage/add');
-    $sharemessage = array(
-      'label' => 'ShareMessage Test AddThis',
-      'id' => 'sharemessage_test_addthis_label',
-      'plugin' => 'addthis',
-      'title' => 'AddThis test',
-    );
-    $this->drupalPostForm(NULL, $sharemessage, t('Save'));
-
-    // Assert that the initial settings are saved correctly.
-    $this->drupalGet('sharemessage-test/sharemessage_test_addthis_label');
-    $this->assertShareButtons($sharemessage, $default_settings['default_icon_style'], TRUE);
-    $this->assertRaw('<a class="addthis_button_facebook">');
-    $this->assertRaw('<a class="addthis_button_facebook_like">');
-    $this->assertNoRaw('<a class="addthis_button_compact">');
-
-    // Set new AddThis settings.
-    $this->drupalGet('admin/config/services/sharemessage/addthis-settings');
-    $default_settings = array(
-      'default_services[]' => array(
-        'facebook',
-        'linkedin',
-        'twitter'
-      ),
-      'default_additional_services' => TRUE,
-      'default_icon_style' => 'addthis_32x32_style',
-    );
-    $this->drupalPostForm(NULL, $default_settings, t('Save configuration'));
-
-    // Check that the saving of the new AddThis settings works correctly.
-    $this->drupalGet('sharemessage-test/sharemessage_test_addthis_label');
-    $this->assertShareButtons($sharemessage, $default_settings['default_icon_style'], TRUE);
-    $this->assertRaw('<a class="addthis_button_facebook">');
-    $this->assertNoRaw('<a class="addthis_button_facebook_like">');
-    $this->assertRaw('<a class="addthis_button_linkedin">');
-    $this->assertRaw('<a class="addthis_button_twitter">');
-    $this->assertRaw('<a class="addthis_button_compact">');
   }
 
 }
