@@ -3,6 +3,7 @@
 namespace Drupal\sharemessage\Tests;
 
 use Drupal\Component\Render\FormattableMarkup;
+use Drupal\Component\Utility\SafeMarkup;
 use Drupal\Component\Utility\Xss;
 use Drupal\simpletest\WebTestBase;
 
@@ -13,7 +14,7 @@ abstract class ShareMessageTestBase extends WebTestBase {
    *
    * @var array
    */
-  public static $modules = array('sharemessage', 'sharemessage_test', 'block', 'filter', 'file');
+  public static $modules = ['sharemessage', 'sharemessage_test', 'block', 'filter', 'file'];
 
   /**
    * Permissions for the admin user.
@@ -43,6 +44,27 @@ abstract class ShareMessageTestBase extends WebTestBase {
     $this->drupalLogin($this->adminUser);
 
     $this->drupalPlaceBlock('page_title_block');
+  }
+
+  /**
+   * Prepares the renderer array for a specific context, sets its raw content.
+   *
+   * @param string $entity_type
+   *   The entity type for this view builder.
+   * @param \Drupal\Core\Entity\EntityInterface $entity
+   *   The entity to render.
+   * @param string $view_mode
+   *   (optional) The view mode that should be used to render the entity.
+   *   Defaults to 'teaser'.
+   */
+  protected function setEntityRawContent($entity_type, $entity, $view_mode = 'teaser') {
+    // The structured array describing the data to be rendered.
+    $elements = \Drupal::entityTypeManager()
+      ->getViewBuilder($entity_type)
+      ->view($entity, $view_mode);
+
+    $content = \Drupal::service('renderer')->renderRoot($elements);
+    $this->setRawContent($content);
   }
 
   /**
@@ -134,10 +156,10 @@ abstract class ShareMessageTestBase extends WebTestBase {
 
     if (!$message) {
       if (!$not_exists) {
-        $message = new FormattableMarkup('Icon style "@raw_html_icon_style" found.', array('@raw_html_icon_style' => $raw_html_icon_style));
+        $message = new FormattableMarkup('Icon style "@raw_html_icon_style" found.', ['@raw_html_icon_style' => $raw_html_icon_style]);
       }
       else {
-        $message = new FormattableMarkup('Icon style "@raw_html_icon_style" not found.', array('@raw_html_icon_style' => $raw_html_icon_style));
+        $message = new FormattableMarkup('Icon style "@raw_html_icon_style" not found.', ['@raw_html_icon_style' => $raw_html_icon_style]);
       }
     }
 
@@ -213,10 +235,10 @@ abstract class ShareMessageTestBase extends WebTestBase {
 
     if (!$message) {
       if (!$not_exists) {
-        $message = new FormattableMarkup('OG tag "@meta_tag" found.', array('@meta_tag' => $meta_tag));
+        $message = new FormattableMarkup('OG tag "@meta_tag" found.', ['@meta_tag' => $meta_tag]);
       }
       else {
-        $message = new FormattableMarkup('OG tag "@meta_tag" not found or has not the expected content.', array('@meta_tag' => $meta_tag));
+        $message = new FormattableMarkup('OG tag "@meta_tag" not found or has not the expected content.', ['@meta_tag' => $meta_tag]);
       }
     }
 
@@ -226,5 +248,61 @@ abstract class ShareMessageTestBase extends WebTestBase {
     else {
       return $this->assertRaw($meta_tag, $message);
     }
+  }
+
+  /**
+   * Asserts that a checkbox field in the current page is checked.
+   *
+   * Because of ajax result, the IDs are not reliable, thus we need to identify
+   * the fields based on the names.
+   * @todo remove this method once #2735005 is committed.
+   *
+   * @param string $name
+   *   The field's name to assert.
+   * @param string $message
+   *   (optional) A message to display with the assertion. Do not translate
+   *   messages: use \Drupal\Component\Utility\SafeMarkup::format() to embed
+   *   variables in the message text, not t(). If left blank, a default message
+   *   will be displayed.
+   * @param string $group
+   *   (optional) The group this message is in, which is displayed in a column
+   *   in test output. Use 'Debug' to indicate this is debugging output. Do not
+   *   translate this string. Defaults to 'Browser'; most tests do not override
+   *   this default.
+   *
+   * @return bool
+   *   TRUE on pass, FALSE on fail.
+   */
+  protected function assertFieldCheckedByName($name, $message = '', $group = 'Browser') {
+    $elements = $this->xpath('//input[@name=:name]', [':name' => $name]);
+    return $this->assertTrue(isset($elements[0]) && !empty($elements[0]['checked']), $message ? $message : SafeMarkup::format('Checkbox field @name is checked.', ['@name' => $name]), $group);
+  }
+
+  /**
+   * Asserts that a checkbox field in the current page is not checked.
+   *
+   * Because of ajax result, the IDs are not reliable, thus we need to identify
+   * the fields based on the names.
+   * @todo remove this method once #2735005 is committed.
+   *
+   * @param string $name
+   *   The field's name to assert.
+   * @param string $message
+   *   (optional) A message to display with the assertion. Do not translate
+   *   messages: use \Drupal\Component\Utility\SafeMarkup::format() to embed
+   *   variables in the message text, not t(). If left blank, a default message
+   *   will be displayed.
+   * @param string $group
+   *   (optional) The group this message is in, which is displayed in a column
+   *   in test output. Use 'Debug' to indicate this is debugging output. Do not
+   *   translate this string. Defaults to 'Browser'; most tests do not override
+   *   this default.
+   *
+   * @return bool
+   *   TRUE on pass, FALSE on fail.
+   */
+  protected function assertNoFieldCheckedByName($name, $message = '', $group = 'Browser') {
+    $elements = $this->xpath('//input[@name=:name]', [':name' => $name]);
+    return $this->assertTrue(isset($elements[0]) && empty($elements[0]['checked']), $message ? $message : SafeMarkup::format('Checkbox field @name is not checked.', ['@name' => $name]), $group);
   }
 }
