@@ -43,8 +43,8 @@ class MigrateUpgrade7Test extends MigrateUpgradeTestBase {
       'configurable_language' => 4,
       'contact_form' => 3,
       'editor' => 2,
-      'field_config' => 48,
-      'field_storage_config' => 36,
+      'field_config' => 49,
+      'field_storage_config' => 37,
       'file' => 1,
       'filter_format' => 7,
       'image_style' => 6,
