@@ -18,11 +18,6 @@ $connection->insert('key_value')
     'name' => 'rest',
     'value' => 'i:8201;',
   ])
-  ->fields([
-    'collection' => 'system.schema',
-    'name' => 'serialization',
-    'value' => 'i:8201;',
-  ])
   ->execute();
 
 // Update core.extension.
@@ -33,8 +28,8 @@ $extensions = $connection->select('config')
   ->execute()
   ->fetchField();
 $extensions = unserialize($extensions);
-$extensions['module']['rest'] = 8201;
-$extensions['module']['serialization'] = 8201;
+$extensions['module']['rest'] = 0;
+$extensions['module']['serialization'] = 0;
 $connection->update('config')
   ->fields([
     'data' => serialize($extensions),
