@@ -17,11 +17,6 @@ $connection->insert('key_value')
     'name' => 'rest',
     'value' => 'i:8000;',
   ])
-  ->fields([
-    'collection' => 'system.schema',
-    'name' => 'serialization',
-    'value' => 'i:8000;',
-  ])
   ->execute();
 
 // Update core.extension.
@@ -32,8 +27,8 @@ $extensions = $connection->select('config')
   ->execute()
   ->fetchField();
 $extensions = unserialize($extensions);
-$extensions['module']['rest'] = 8000;
-$extensions['module']['serialization'] = 8000;
+$extensions['module']['rest'] = 0;
+$extensions['module']['serialization'] = 0;
 $connection->update('config')
   ->fields([
     'data' => serialize($extensions),
@@ -52,7 +47,7 @@ $config = [
       ],
     ],
   ],
-  'link_domain' => '~',
+  'link_domain' => NULL,
 ];
 $data = $connection->insert('config')
   ->fields([
