@@ -1172,7 +1172,7 @@ abstract class ArgumentPluginBase extends HandlerBase implements CacheableDepend
    * Moves argument options into their place.
    *
    * When configuring the default argument behavior, almost each of the radio
-   * buttons has its own fieldset shown bellow it when the radio button is
+   * buttons has its own fieldset shown below it when the radio button is
    * clicked. That fieldset is created through a custom form process callback.
    * Each element that has #argument_option defined and pointing to a default
    * behavior gets moved to the appropriate fieldset.
