@@ -175,7 +175,7 @@ abstract class BrowserTestBase extends \PHPUnit_Framework_TestCase {
    *
    * @see \Drupal\Tests\BrowserTestBase::installDrupal()
    */
-  public static $modules = [];
+  protected static $modules = [];
 
   /**
    * An array of config object names that are excluded from schema checking.
