@@ -71,7 +71,7 @@ class SystemStateEdit extends FormBase {
     $form['value'] = array(
       '#type' => 'item',
       '#title' => $this->t('Current value for %name', array('%name' => $state_name)),
-      '#markup' => kprint_r($old_value, TRUE),
+      '#markup' => kpr($old_value, TRUE),
     );
 
     $transport = 'plain';
