@@ -4,12 +4,13 @@ namespace Drupal\webprofiler\Cache;
 
 use Drupal\Core\Cache\Cache;
 use Drupal\Core\Cache\CacheBackendInterface;
+use Drupal\Core\Cache\CacheTagsInvalidatorInterface;
 use Drupal\webprofiler\DataCollector\CacheDataCollector;
 
 /**
  * Wraps an existing cache backend to track calls to the cache backend.
  */
-class CacheBackendWrapper implements CacheBackendInterface {
+class CacheBackendWrapper implements CacheBackendInterface, CacheTagsInvalidatorInterface {
 
   /**
    * The data collector to register the calls.
@@ -124,13 +125,6 @@ class CacheBackendWrapper implements CacheBackendInterface {
   /**
    * {@inheritdoc}
    */
-  public function deleteTags(array $tags) {
-    return $this->cacheBackend->deleteTags($tags);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
   public function deleteAll() {
     return $this->cacheBackend->deleteAll();
   }
@@ -153,7 +147,9 @@ class CacheBackendWrapper implements CacheBackendInterface {
    * {@inheritdoc}
    */
   public function invalidateTags(array $tags) {
-    return $this->cacheBackend->invalidateTags($tags);
+    if ($this->cacheBackend instanceof CacheTagsInvalidatorInterface) {
+      $this->cacheBackend->invalidateTags($tags);
+    }
   }
 
   /**
@@ -176,4 +172,5 @@ class CacheBackendWrapper implements CacheBackendInterface {
   public function removeBin() {
     return $this->cacheBackend->removeBin();
   }
+
 }
