@@ -9,15 +9,24 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Drupal\Console\Command\ContainerAwareCommand;
+use Drupal\Console\Core\Command\Shared\ContainerAwareCommandTrait;
+use Symfony\Component\Console\Command\Command;
 use Symfony\Component\DomCrawler\Crawler;
 use Symfony\Component\Process\Process;
 use Symfony\Component\Yaml\Yaml;
+use Drupal\Console\Annotations\DrupalCommand;
 
 /**
  * Class BenchmarkCommand
+ *
+ * @DrupalCommand (
+ *     extension="webprofiler",
+ *     extensionType="module"
+ * )
  */
-class BenchmarkCommand extends ContainerAwareCommand {
+class BenchmarkCommand extends Command {
+
+  use ContainerAwareCommandTrait;
 
   /**
    * {@inheritdoc}
@@ -56,7 +65,7 @@ class BenchmarkCommand extends ContainerAwareCommand {
     }
 
     /** @var \Drupal\Core\Http\Client $client */
-    $client = $this->getContainer()->get('http_client');
+    $client = $this->container->get('http_client');
 
     $progress = new ProgressBar($output, $runs + $steps);
     $progress->setFormat(' %current%/%max% [%bar%] %percent:3s%% %message%');
@@ -152,7 +161,7 @@ class BenchmarkCommand extends ContainerAwareCommand {
     $token = $response->getHeader('X-Debug-Token');
 
     /** @var \Drupal\webprofiler\Profiler\Profiler $profiler */
-    $profiler = $this->getContainer()->get('profiler');
+    $profiler = $this->container->get('profiler');
 
     /** @var \Symfony\Component\HttpKernel\Profiler\Profile $profile */
     $profile = $profiler->loadProfile($token);
