@@ -6,23 +6,31 @@ use Drupal\Core\State\StateInterface;
 use Drupal\webprofiler\DataCollector\StateDataCollector;
 
 /**
- * Class StateWrapper
+ * Class StateWrapper.
  */
 class StateWrapper implements StateInterface {
 
   /**
+   * The system state.
+   *
    * @var \Drupal\Core\State\StateInterface
    */
   private $state;
 
   /**
+   * The state data collector.
+   *
    * @var \Drupal\webprofiler\DataCollector\StateDataCollector
    */
   private $dataCollector;
 
   /**
+   * StateWrapper constructor.
+   *
    * @param \Drupal\Core\State\StateInterface $state
+   *   The system state.
    * @param \Drupal\webprofiler\DataCollector\StateDataCollector $dataCollector
+   *   The state data collector.
    */
   public function __construct(StateInterface $state, StateDataCollector $dataCollector) {
     $this->state = $state;
@@ -83,4 +91,20 @@ class StateWrapper implements StateInterface {
   public function resetCache() {
     $this->state->resetCache();
   }
+
+  /**
+   * Passes through all non-tracked calls onto the decorated object.
+   *
+   * @param string $method
+   *   The called method.
+   * @param mixed $args
+   *   The passed in arguments.
+   *
+   * @return mixed
+   *   The return argument of the call.
+   */
+  public function __call($method, $args) {
+    return call_user_func_array([$this->state, $method], $args);
+  }
+
 }
